+++ src/main/java/com/errand/services/TaskService.java
@@ -18,6 +18,6 @@ public interface TaskService {
 
     Task saveTask(TaskDto taskDto);
 
-    List<TaskDto> findTaskByServiceProvider(Integer id);
+    List<TaskDto> findTaskByServiceProvider(Long id);
 
 }
